Feature/confi application

See merge request cordytech-lab/ms-operaciones!3

+++ src/main/java/com/cordytech/ms_operaciones/model/Boleta.java
@@ -1,14 +1,24 @@
 package com.cordytech.ms_operaciones.model;
 
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import java.time.LocalDateTime;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Table(name = "boletas")
@@ -28,7 +38,7 @@ public class Boleta {
     private LocalDateTime fechaEmision;
 
     @Column(name = "monto", nullable = false, precision = 19, scale = 2)
-    private Double monto;
+    private BigDecimal monto;
 
     @Column(name = "estado", nullable = false)
     @Enumerated(EnumType.STRING)
